Add analysis for drainage areas
This takes an input AoI, finds its centroid, finds a HUC-12 that
contains the centroid, and runs GWLF-E on it. Ultimatley, we will
scale the GWLF-E results to the input AoI, but for that we need
the Drexel / ANS API. So this is not complete until that API
has been integrated.