Implement Drainage Area Analyze View
When the area of interest is a Drainage Area,
we show an additional Analyze Tab, which shoes contents
of the GWLF-E Runoff and Water Quality views.

The Drainage Area Result View is made quite akin
to the Analyze Result View, but had to be seprate
because of its custom CSV Export handling.