Unversion Firefox, Upgrade PostGIS to 3.3
We use Firefox for JS testing. Every 10 versions we have to
update the version string. Firefox has been stable for the
last few years, and the tests have not broken. By unversioning
it we will get the latest version every time, which should be
fine given the stability over the last few years.

PostGIS 3.3 is needed because a step in provisioning requests
ST_TriangulatePolygon, which is only available in PostGIS 3.3.
It is unclear what is requesting that. This does not have an
impact on us, because we're not using any feature that is
deprecated in PostGIS 3.3.