Clear out unaligned NLCD19 GWLF-E results
Adds a migration that clears out all stored results for GWLF-E
projects created on or after 2022-01-17, which is when 1.33.0
was released with incorrectly aligned NLCD19 layers, which had
also been made the default. Thus, every project made after then
had slighly incorrect results.

This migration clears out those bad results. The next time that
project is loaded in the UI, all its results will be recalculated.
This allows us to rebuild the results over time, and not overload
the system all at once.