Upgrade Redis instance in production
We've been on Redis 2 in production for a long time, and it is
being EOL'd. We've been using Redis 5 in development since 5bf37aa
in August 2021. Thus, with this experience, we upgrade the live
instance to Redis 5 as well.

These files are updated for the record, but the actual  upgrade
was performed via the UI. A full record of actions taken can be
seen here: https://github.com/WikiWatershed/model-my-watershed/issues/3576#issuecomment-1384390400