Use all environment variables for mmw-geoprocessing
Previously, only a handful of environment variables were supplied
to the mmw-geoprocessing service. This was confusing, as more
relevant variables were added to mmw.d/env but were not taking
effect.

By loading all the environment variables for mmw-geoprocessing,
just as we do for celery, we make the environment variables
more universal and consistently applied.