Switch RWD to use direct Docker service
The docker_compose module is no longer supported by Ansible, and
switching to the docker_compose_v2 module required updating to a
new version of Docker which does not work with the v1 Docker Image
format of the RWD image.

By using a service with inline docker cli commands, we replicate
the settings in the docker_compose block, and also bring it more
in line with the Geoprocessing service.