Add Celery task to query STAC and generate histogram
This is based on preparatory work done in https://github.com/rajadain/mmw-io-10m-lulc-summary/blob/main/main.py

We create a Celery task that takes a STAC URL, collection, asset,
and optional filters, and generates a histogram from it.

There is an additional method to format this into an MMW Geoprocessing
like shape, which allows for compatibility and modularity.