Mark DRB Projections as Lazy
Marking analyses as lazy makes them not execute until they are
selected by users. Given that the Drexel / ANS endpoint that
powers these anlayses has been broken for a long time, it does
not serve to have them execute by default for every run. This
edit makes it so that they don't.