Disable CloudFront Caching for tiles
Our design uses S3 for tile caching. The front-end first hits
S3, and if it doesn't find a tile there, it hits the Tile Server.
This is handled through a custom redirection in S3 which handles
404s with a 302 to the Tile Server. Unfortunately, this 302 gets
cached by CloudFront, making subsequent requests to the now
available tile in S3 also go to the Tile Server, resulting in a
slower UX. By disabling caching at the CloudFront layer, we
ensure that all S3 requests go straight to S3 in every case, and
we make minimal use of the Tile Server.

This may potentially increase hosting costs as S3 will now be
more utilized, but will lead to a much faster UX. These costs
may be addressed at a later stage if they become too onerous.