Ingest TDX Hydro Streams
In practice, given the dataset is so big, it may be better
to download it first, perhaps using a specialized tool like
s5cmd, then to extract it in parallel using pigz which would
be faster than the default gzip, then import it manually.

Including this here for completion.

The Services VM is also increased to using 128GB of space
which is now necessary with this new dataset.