Do not cache tiles in CloudFront, only S3
If a tile is not found in S3, we redirect to the tiler VM
to generate a new tile. Unfortunately, this redirect is
cached by CloudFront, causing overuse of the tiler and
slower performance overall, despite the cached tile being
available in S3.

Since we cannot configure CloudFront to not cache redirects,
we turn off CloudFront caching completely. This will increase
the S3 usage, but also makes the UX much faster.