Add back-end for analyzing global streams
Adds a new URL and View for Global Stream Analysis, but
reuses the same Task and Calc under the hood as the NHD
streams. This is because we're doing the same calculation
and the tables have the same column names.

In the case of global streams, we are not doing the Land
Use Intersection which informs the Ag Length value, so
that is replaced with None for now.