Use albers projection based on AoI location
Since this calculation now runs on TDX Hydro, a global dataset,
we can no longer reproject the results into 5070, which only
works for North America. The projection calculation is moved to
calcs.py where it can be reused, and now called both in the
Land Analysis and the Streams Analysis.