Move TiTiler parameters into environment variables
Gets rid of the hard-coded values for the TiTiler URL and layer UUID mappings
and puts the values in environment variables instead. The layer mapping is
complex and gets encoded in the variable as a comma-separated list of strings
with the values (layer, year, and UUID) separated by '__'.
The coded parses that into a {layer: { year: UUID }} object.

Note that the setup this is replacing didn't have the 'layer' value, it just
assumed there was only one. Currently there's only one, but including it in
the config will make it easier to add more in the future.