Add IO LULC layers to LAYER_GROUPS setting
Adds the new global land cover layers, from 2017 to 2023, and removes the
one from the previous proof-of-concept hybrid tiler setup.
To be able to make the URL dynamic by environment (but without creating
an import loop between layer_settings.py and base.py), this moves the
MMW_TILER_HOST definition into here and adds it to the list of variables
imported into `base.py`.