Validate AoI against perimeter if specified
Now that we have standardized perimeters, we add a check deep
in the AoISerializer to check if the AoI is contained within the
perimeter, if specified. If no perimeter is specified, then the
check will be skipped.

While this requires a lot of parameter drilling, getting it it
deep into MultiPolygonGeoJsonField.to_internal_value allows us
to use the GEOSGeometry object made of the AoI, without having
to reserialize the returned JSON into a new geometry object,
thus making it faster.

The universal CONUS check is removed, since now we accept AoIs
that are outside CONUS, in some global cases.