Disable CONUS-specific draw tools when outside
The Select Boundary draw tool is only avialable within CONUS,
since all our boundaries are CONUS only, so that gets disabled
completely when the CONUS is outside the map viewport.

For RWD, the NHD and DRB options are disabled when outside
CONUS. There will be an upcoming global RWD option that will
allow users to draw boundaries anywhere in the world.

Closes #3650
Closes #3651