Hide disabled layers in selector
Previously, when the viewport was outside the layer's
perimeter, we would disabled (gray out) the layer in
the selector. Unfortunately, due to the large number
of layers and the hidden scrollbar in the layer selector,
when the top layers are grayed out, it is not obvious to
users that other layers are available.

By hiding the disabled layers completely, we make it so
that only the available layers are shown in the selector,
which makes it easy to see which options are available
at any given point on the map.