Add back-end for delineating global watersheds
The actual SQL query is very fast, but we still wrap it
in a Celery job so that it is logged and recorded, and
is consistent with the other APIs.

In case no matching basins are found, the returning
feature would have a None geometry.

For details on this algorithm, please see:
https://github.com/WikiWatershed/global-hydrography/?tab=readme-ov-file#modified-nested-set-index