Refactor: Use single watershed/ API
Previously, we had added a new global-watershed/ API endpoint
for TDX Basins. This was hard to discover as alphabetically
it was distant from the watershed/ API that already existed in
the API Documentation. Since the output format is almost identical
to the existing watershed/ endpoint, by combining the two into
a single endpoint, with TDX being just one of the sources, we
simplify and unify our API.