Set max bounds to prevent overflow errors
Previously, with no max bounds set, a user could
pan west from the US and end up in Australia, in
a way that exceeds the nominal EPSG:4326 bounds.
This resulted in GeoJSON that needed to be fixed
to a normal range within (-180,180) before using
to intersect with other ranges.

By limitin the map to within nominal bounds, we
ensure that all polygons generated in the map
will be correct.