Apply area-derived modifications first, then user
Previously, we would apply all user-specified modifications
first, then apply the area-derived ones. This was done in #3151,
and was a mistake, since the area-derived modifications would
override the user-specified ones.

This was not caught until users tried to update Normal Septic
Systems in #3694, which is a field that is area-derived.

By making this change, we ensure that all area-derived modifications
are applied first, and then user-specified are applied last.
This ensures that all user-specified modifications do take
effect in the final calculations.