Refactor frontend PA perimeter check to match analysis backend
The backend point source endpoint relies on the MultiPolygonGeoJsonField serializer to validate an area of interest is within a perimeter. Similarly, the frontend conditionally displays datasource choices and sets the default path parameter by geography, validating via a custom frontend utility fn. The backend utilizes the PostGIS contains method which is more strict than our frontend custom check, so this discrepency can lead to a datasource choice that will fail on request. This commit creates a new, more strict perimeter check for PA that matches the results of the backend validation for areas of interest that are not completely within state bounds.