Move weather simulation eligibility into project model
Previously, we would check if a project is eligible for
some weather simulations based on its AoI. This was a proxy
to check if the two nearest weather stations for a project
would be one of the Recent or Future weather datasets.

Now we shift to checking that directly. Since this is an
expensive check requiring database checks, geospatial checks,
and S3 checks, we compute this once and save it to the
project.

This is done in a new JSON field called weather_simulations,
whose contents mimic WeatherType.Simulation constant from
the front-end. Each group in this JSON list has a disabled
value, which is True by default. For certain cases, we will
turn that False, thus enabling it in the UI.

All existing projects are seeded with the correct value
based on its gis_data -> WeatherStations list. All GWLFE
projects are initialized with both datasets disabled. Then,
for GWLFE projects whose both weather stations are present
in Recent weather dataset, we turn enable Recent. Same for
Future. Finally, those that have both weather stations in
the common intersection of these lists have both enabled.