Add preset support to entry sections
For point sources, we need a way to populate the overrides
with presets. This is currently available in the Entry aka
Settings modal for GWLFE projects, in the Waste Water
section. To keep the preset selection near those entries,
we add the ability to specify presets for each kind of
entry.

This expects a list of presets, and a presetUrl() function
to be passed in, which will be used to fetch values and
populate a modification. The preset is also saved to the
modification, so we will remember the selection for when
the user comes back to this project.