Fix weather simulation check on create
The AoI function requires a geometry, not a dict. This
conversion happens in the serializer, so we move the
check to after serialization. We don't need to check
for AoI anymore since any project must have it as part
of validation.

We also don't catch the error, and instead throw it
instead, making a failure louder and more prominent.

We allow specification of this dictionary, necessary
for tests which cannot access the weather stations table.